web/settings: add a reusable `SettingsInput` component
